Update patch set 1
Patch Set 1:

Well just so you know GPS on mi8 checks for the Product I'd and throws up an error in logs as we have it now. We may change that part of your commit at some point anyway. 

If you could find a working mi8 build fingerprint that would be appreciated ;)

Patch-set: 1