Update patch set 7
Patch Set 7:

I checked stock configs and there is no doze light sensor specified. With these settings active and the other related ones from blueline I can see it tries to enable the lightsensor in logs but its not effective as it does not respond in anyway to light changes.

Stock configs have a different doze brightness of 17 instead of 7 but I dont think there is any reason to change from this. displayBlanksAfterDoze is set to true in stock and this seems to fix the AOD bug when receiving a notification. 

Doze display brightness in buckets is also false in stock.

Patch-set: 7
Reviewer: Gerrit User 20225 <20225@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>