Update patch set 5
Patch Set 5:

(1 comment)

The proper fix would be to edit /vendor/etc/gps.conf and enable injection there. I am fine shipping this in /system/etc/ while we are using the prebuilt vendor images, but once we switch we should do it properly. As the name suggests, this is just a debug conf file used to override carrier settings.

To be sure that this is exactly what fixes it, are you able to do one final test? Remove the file from /system/etc/ and edit /vendor/etc/gps.conf. Let me know if it works.

Nonetheless, as I told you already, it's better to import the config into the device tree rather than making it a blob so that we are able to edit in the future, if needed.

Patch-set: 5
Reviewer: Gerrit User 13648 <13648@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>