Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I have removed white led max brightness limitation due to the comments that I left in https://review.lineageos.org/c/LineageOS/android_device_xiaomi_sdm845-common/+/248348.

I thought I'd replied to that. Oops. Anyway why would we limit one led and not the other? I assume that allowing the LEDs too much current will burn them out quicker? Battery led stays on while charging so if so will be running max current during that time. To limit the brightness for notification led and battery led brightness I have to find an overlay for that? Tbh I don't really see the benefit of doing it that way or the problem of doing it this way?

Patch-set: 4
Reviewer: Gerrit User 20225 <20225@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>