Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > I have removed white led max brightness limitation due to the comments that I left in https://review.lineageos.org/c/LineageOS/android_device_xiaomi_sdm845-common/+/248348.
> 
> I thought I'd replied to that. Oops. Anyway why would we limit one led and not the other? I assume that allowing the LEDs too much current will burn them out quicker? Battery led stays on while charging so if so will be running max current during that time. To limit the brightness for notification led and battery led brightness I have to find an overlay for that? Tbh I don't really see the benefit of doing it that way or the problem of doing it this way?

The problem of doing it this way is that you're limiting the max brightness. What if one wants to have it as brighter than stock? The current range for led brightness is 0-255 and you're suggesting to limit it to 0-6. That renders led brightness settings useless if such range is used.

Patch-set: 4
Reviewer: Gerrit User 13648 <13648@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>